fix(kynos): correct the integer formats and the null union
`u32` claimed `format: int32` alongside `maximum: 4294967295`. The OAS
Format Registry defines `int32` and `int64` as *signed*, so the schema
contradicted itself and a generator honouring the format would emit a type
its own stated maximum overflows. The bounds stay exact; the format widens to
`int64`.

`Option<()>` produced `type: [null, null]`, and `Option<Option<T>>` an
`anyOf` with two identical branches. A JSON Schema type union's members must
be unique, so a schema that already admits `null` widens to itself.

`MapKey` was a bare marker whose documented invariant -- that the key
describes itself as a string -- nothing checked, so `impl MapKey for u32`
compiled and produced `propertyNames: {"type": "integer"}`, matching no
object that can exist. It now supplies *constraints* rather than a schema,
and the map builds `propertyNames` as a string schema plus those. String-ness
is true by construction, and the trait stays open to the string-shaped
newtype its own documentation recommends. Sealing it would have closed that
door instead.

The schema helpers had no tests at all, because every composite
implementation routes through `Registry::resolve`, which is still `todo!()`.
`nullable` takes a schema rather than a registry, so it is testable today --
which is how both of the above were confirmed.

Also: the handler diagnostic now mentions `Send`. A future holding an `Rc`
across an await failed with four notes about `Describe`, `IntoResponse` and
`Responses`, none of which was the problem.

+++ crates/kynos/src/schema/impls/primitive.rs
@@ -60,12 +60,18 @@ macro_rules! unsigned {
 }
 
 signed!(i8 => "int32", i16 => "int32", i32 => "int32");
-unsigned!(u8 => "int32", u16 => "int32", u32 => "int32");
+
+// `int32` and `int64` are *signed* in the OAS Format Registry, so a `u32` does
+// not fit `int32` — a generator honouring the format would emit a type its own
+// maximum overflows. The bounds stay exact; the format widens.
+unsigned!(u8 => "int32", u16 => "int32", u32 => "int64");
 
 // `i64::MAX` and `u64::MAX` are not representable in an `f64`, and JSON Schema
 // bounds are numbers. Emitting a rounded bound would forbid values the type
 // accepts, or accept values it does not, so the width is left to the format —
-// which is the only honest thing the vocabulary can say here.
+// which is the only honest thing the vocabulary can say here. `u64` keeps
+// `int64` for the same reason it keeps no maximum: nothing narrower is true,
+// and its `minimum: 0` is.
 impl Schema for i64 {
     fn schema(_registry: &mut Registry) -> OpenApiSchema {
         formatted(SchemaType::Integer, "int64")
